refactor: LottoStore 기능을 LottoMachine 으로 이동
