+++ src/main/java/lotto/domain/LottoMachine.java
@@ -11,8 +11,9 @@ import java.util.stream.IntStream;
 
 import static lotto.util.CollectionUtils.*;
 
-class LottoMachine {
+public class LottoMachine {
 
+	public static final int PRICE = 1000;
 	private static final int FROM_INDEX = 0;
 	private static final int LOTTO_NUMBER_COUNT = 6;
 	private static final int LOTTO_NUMBER_LOWER_BOUND = 1;
@@ -23,6 +24,29 @@ class LottoMachine {
 		return IntStream.rangeClosed(LOTTO_NUMBER_LOWER_BOUND, LOTTO_NUMBER_UPPER_BOUND)
 				.mapToObj(LottoNumber::of)
 				.collect(Collectors.collectingAndThen(Collectors.toList(), Collections::unmodifiableList));
+	}
+
+	public UserLotto buy(LottoBuyingRequest lottoBuyingRequest) {
+		return new UserLotto(Collections.unmodifiableList(generateLottoTickets(lottoBuyingRequest)));
+	}
+
+	private List<LottoTicket> generateLottoTickets(LottoBuyingRequest lottoBuyingRequest) {
+		List<LottoTicket> lottoTicketList = new ArrayList<>(generateManualLottoTickets(lottoBuyingRequest));
+		lottoTicketList.addAll(generateAutoLottoIfRequestExist(lottoBuyingRequest.autoLottoCount()));
+
+		return lottoTicketList;
+	}
+
+	private List<LottoTicket> generateManualLottoTickets(LottoBuyingRequest lottoBuyingRequest) {
+		return manual(lottoBuyingRequest.manualLottoNumberStrings());
+	}
+
+	List<LottoTicket> manual(List<String> numberStrings) {
+		return transform(numberStrings, new ArrayList<>(), LottoTicketConverter::convert);
+	}
+
+	private List<LottoTicket> generateAutoLottoIfRequestExist(int count) {
+		return new ArrayList<>(generate(count));
 	}
 
 	List<LottoTicket> generate(int count) {
@@ -41,9 +65,5 @@ class LottoMachine {
 		List<LottoNumber> shuffledList = tempLottoNumberList.subList(FROM_INDEX, LOTTO_NUMBER_COUNT);
 
 		return new LottoTicket(Collections.unmodifiableSet(new HashSet<>(shuffledList)));
-	}
-
-	public List<LottoTicket> manual(List<String> numberStrings) {
-		return transform(numberStrings, new ArrayList<>(), LottoTicketConverter::convert);
 	}
 }
