+++ src/test/java/lotto/domain/LottoReportTest.java
@@ -6,8 +6,8 @@ import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 
+import static lotto.domain.LottoMachine.*;
 import static lotto.domain.LottoRank.*;
-import static lotto.domain.LottoStore.*;
 import static org.assertj.core.api.Assertions.*;
 
 class LottoReportTest {
